# BEAST-Server/res/content/api-docs/OpenAPI.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "D-FARM API",
    "description" : "D-FARM API",
    "contact" : {
      "name" : "CodeInspect Support Team",
      "url" : "https://codeinspect.de",
      "email" : "helpdesk@codeinspect.de"
    },
    "version" : "0.2"
  },
  "security" : [ {
    "APIKey" : [ ]
  } ],
...
  "paths" : {
    "/api/devices/{devid}/apps/{appid}" : {
      "get" : {
        "tags" : [ "Apps" ],
        "summary" : "Shows information about a specific app",
        "description" : "Shows information about a specific app",
        "operationId" : "getInstalledAppInfo",
        "parameters" : [ {
          "name" : "devid",
          "in" : "path",
          "description" : "The id of device",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "appid",
          "in" : "path",
          "description" : "The id of app",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AbstractApp"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "400" : {
            "description" : "Wrong input",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "403" : {
            "description" : "Access denied",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Apps" ],
        "summary" : "Uninstalls an app",
        "description" : "Uninstalls an app",
        "operationId" : "uninstallApplication",
        "parameters" : [ {
          "name" : "devid",
          "in" : "path",
          "description" : "The id of device",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "appid",
          "in" : "path",
          "description" : "The id of app",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AbstractApp"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "400" : {
            "description" : "Wrong input",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "403" : {
            "description" : "Access denied",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/devices/{devid}/contacts" : {
      "get" : {
        "tags" : [ "Contacts" ],
        "summary" : "Lists contacts",
        "description" : "Lists contacts",
        "operationId" : "getContacts",
        "parameters" : [ {
          "name" : "devid",
          "in" : "path",
          "description" : "The id of device",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Contact"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "400" : {
            "description" : "Wrong input",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "403" : {
            "description" : "Access denied",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Contacts" ],
        "summary" : "Adds or updates a contact",
        "description" : "Adds or uploads a contact",
        "operationId" : "addContact",
        "parameters" : [ {
          "name" : "devid",
          "in" : "path",
          "description" : "The id of device",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Contact"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AbstractApp"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "400" : {
            "description" : "Wrong input",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "403" : {
            "description" : "Access denied",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{devid}/contacts/{id}" : {
      "post" : {
        "tags" : [ "Contacts" ],
        "summary" : "Deletes a contact",
        "description" : "Deletes a contact",
        "operationId" : "uninstallApplication_1",
        "parameters" : [ {
          "name" : "devid",
          "in" : "path",
          "description" : "The id of device",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "The id of the contact",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "500" : {
            "description" : "Internal error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "400" : {
            "description" : "Wrong input",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "403" : {
            "description" : "Access denied",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "AbstractApp" : {
        "required" : [ "id", "type" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "An identifier for the app"
          },
          "type" : {
            "type" : "string",
            "description" : "The type of app",
            "readOnly" : true
          }
        },
        "description" : "Information about an app",
        "discriminator" : {
          "propertyName" : "type",
          "mapping" : {
            "AndroidApp" : "#/components/schemas/AndroidApp",
            "AbstractApp" : "#/components/schemas/AbstractApp"
          }
        }
      },
      "AndroidApp" : {
        "required" : [ "allowBackup", "id", "systemApp", "type" ],
        "type" : "object",
        "description" : "Android Application information",
        "allOf" : [ {
          "$ref" : "#/components/schemas/AbstractApp"
        }, {
          "type" : "object",
          "properties" : {
            "versionCode" : {
              "type" : "integer",
              "description" : "The version code of the app",
              "format" : "int32",
              "readOnly" : true
            },
            "targetSDK" : {
              "type" : "integer",
              "description" : "The target SDK version of the app",
              "format" : "int32",
              "readOnly" : true
            },
            "versionName" : {
              "type" : "string",
              "description" : "The version name of the app",
              "readOnly" : true
            },
            "codePath" : {
              "type" : "string",
              "description" : "The path where the app resides on the device",
              "readOnly" : true
            },
            "firstDateInstall" : {
              "type" : "string",
              "description" : "The date of first installation",
              "format" : "date-time",
              "readOnly" : true
            },
            "lastUpdateTime" : {
              "type" : "string",
              "description" : "The date of last update",
              "format" : "date-time",
              "readOnly" : true
            },
            "minSDKVersion" : {
              "type" : "integer",
              "description" : "The minimum sdk version",
              "format" : "int32",
              "readOnly" : true
            },
            "installPermissions" : {
              "type" : "array",
              "description" : "Permission requested upon installation",
              "readOnly" : true,
              "items" : {
                "$ref" : "#/components/schemas/AndroidPermission"
              }
            },
            "declaredPermissions" : {
              "type" : "array",
              "description" : "Permission declared in the manifest",
              "readOnly" : true,
              "items" : {
                "$ref" : "#/components/schemas/AndroidPermission"
              }
            },
            "runtimePermissions" : {
              "type" : "array",
              "description" : "Permission requsted at runtime",
              "readOnly" : true,
              "items" : {
                "$ref" : "#/components/schemas/AndroidPermission"
              }
            },
            "systemApp" : {
              "type" : "boolean",
              "description" : "Whether the app is a system app",
              "readOnly" : true
            },
            "allowBackup" : {
              "type" : "boolean",
              "description" : "Whether the backup is allowed",
              "readOnly" : true
            },
            "packageName" : {
              "type" : "string"
            }
          }
        } ]
      },
      "AndroidPermission" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "granted" : {
            "type" : "boolean"
          }
        },
        "description" : "Permission requsted at runtime",
        "readOnly" : true
      },
      "APIException" : {
        "required" : [ "description", "title" ],
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string",
            "description" : "A short description of the error that occured",
            "example" : "Internal Error"
          },
          "description" : {
            "type" : "string",
            "description" : "A more detailed description of the error that occured",
            "example" : "Out of storage space"
          },
          "code" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "description" : "An exception has occurred while processing the request."
      },
...
      "Contact" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "familyName" : {
            "type" : "string"
          },
          "givenName" : {
            "type" : "string"
          },
          "middleName" : {
            "type" : "string"
          },
          "phoneNumbers" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PhoneNumber"
            }
          },
          "emailAddresses" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EmailAddress"
            }
          },
          "notes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Note"
            }
          },
          "websites" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Website"
            }
          },
          "postalAddresses" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PostalAddress"
            }
          },
          "imAddresses" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/IMAddress"
            }
          },
          "timesContacted" : {
            "type" : "integer",
            "format" : "int32"
          },
          "lastContact" : {
            "type" : "string",
            "format" : "date-time"
          },
          "displayName" : {
            "type" : "string"
          }
        }
      },
      "EmailAddress" : {
        "type" : "object",
        "properties" : {
          "address" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "HOME", "MOBILE", "OTHER", "WORK" ]
          },
          "bindings" : {
            "type" : "string"
          }
        }
      },
      "IMAddress" : {
        "type" : "object",
        "properties" : {
          "address" : {
            "type" : "string"
          },
          "protocol" : {
            "type" : "string"
          },
          "bindings" : {
            "type" : "string"
          }
        }
      },
      "Note" : {
        "type" : "object",
        "properties" : {
          "note" : {
            "type" : "string"
          },
          "bindings" : {
            "type" : "string"
          }
        }
      },
      "PhoneNumber" : {
        "type" : "object",
        "properties" : {
          "number" : {
            "type" : "string"
          },
          "bindings" : {
            "type" : "string"
          }
        }
      },
      "PostalAddress" : {
        "type" : "object",
        "properties" : {
          "country" : {
            "type" : "string"
          },
          "region" : {
            "type" : "string"
          },
          "street" : {
            "type" : "string"
          },
          "formatted_address" : {
            "type" : "string"
          },
          "city" : {
            "type" : "string"
          },
          "neighborHood" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "HOME", "OTHER", "WORK" ]
          },
          "postalCode" : {
            "type" : "string"
          },
          "bindings" : {
            "type" : "string"
          }
        }
      },
      "Website" : {
        "type" : "object",
        "properties" : {
          "URL" : {
            "type" : "string"
          },
          "bindings" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "APIKey" : {
        "type" : "apiKey",
        "description" : "Your API key",
        "name" : "APIKey",
        "in" : "header"
      }
    }
  }
}